Adjust icons on profile and home pages
