+++ webapp/src/pages/Home.jsx
@@ -28,10 +28,6 @@ export default function Home() {
   const [status, setStatus] = useState('checking');
 
   const [photoUrl, setPhotoUrl] = useState('');
-  const [globe, setGlobe] = useState(() => {
-    const hour = new Date().getHours();
-    return hour < 12 ? '🌏' : '🌎';
-  });
 
   useEffect(() => {
     ping()
@@ -63,15 +59,6 @@ export default function Home() {
     return () => window.removeEventListener('profilePhotoUpdated', handleUpdate);
   }, []);
 
-  useEffect(() => {
-    const update = () => {
-      const hour = new Date().getHours();
-      setGlobe(hour < 12 ? '🌏' : '🌎');
-    };
-    const id = setInterval(update, 60 * 60 * 1000);
-    update();
-    return () => clearInterval(id);
-  }, []);
 
   return (
 
@@ -80,13 +67,12 @@ export default function Home() {
       <div className="flex flex-col items-center">
 
         {photoUrl && (
-          <div className="relative">
+          <div>
             <img
               src={photoUrl}
               alt="profile"
               className="w-36 h-36 hexagon border-4 border-brand-gold -mt-[20%] mb-3 object-cover"
             />
-            <span className="absolute -right-6 top-1/2 -translate-y-1/2 text-xl">{globe}</span>
           </div>
         )}
